Fix MSVC build errors and warnings
- Wrap std::min in parentheses in profiler.hpp to avoid macro collision.
- Cast ImGui::GetTime() to float in overlay.cpp to fix conversion warning.
- Add _CRT_SECURE_NO_WARNINGS to CMakeLists.txt to suppress CRT safety warnings.

+++ cpp/client/overlay.cpp
@@ -48,7 +48,7 @@ void Overlay::Render() {
         static int frames = 0;
         static float fps = 0;
         frames++;
-        float currentTime = ImGui::GetTime();
+        float currentTime = (float)ImGui::GetTime();
         if (currentTime - lastTime >= 1.0f) {
             fps = (float)frames / (currentTime - lastTime);
             frames = 0;
